requirements.txt
apturl==0.5.2
asgiref==3.3.4
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info==1.0
Django==3.2.5
djangorestframework==3.12.4
gyp==0.1
httplib2==0.18.1
idna==2.10
jeepney==0.6.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.16.0
macaroonbakery==1.3.1
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
psycopg2==2.8.6
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.1
python-debian==0.1.39
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
wadllib==1.3.5
xdg==5
xkit==0.0.0
